requirements.txt
py-cord
PyNaCl
yt-dlp
python-dotenv
openai
torch
numpy
vosk
